fix(build): move shared into server
