+++ .github/workflows/ci.yml
@@ -28,9 +28,8 @@ jobs:
       - run: npm ci
       - run: npx lockfile-lint --type npm --path client/package-lock.json --validate-https --allowed-hosts npm
       - run: npx lockfile-lint --type npm --path server/package-lock.json --validate-https --allowed-hosts npm
-      - run: npx lockfile-lint --type npm --path shared/package-lock.json --validate-https --allowed-hosts npm
       - run: npm run lint-ci
-      - run: npm run --prefix server build
+      - run: npm run --prefix server build -- --noEmit
       - run: npm run --prefix client build
   test:
     name: Test
