+++ client/tsconfig.paths.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "~shared/*": ["../shared/*"]
+      "~shared/*": ["../server/src/shared/*"]
     }
   }
 }
