+++ package.json
@@ -7,17 +7,16 @@
     "start": "npm --prefix server start",
     "build-frontend": "npm run build --prefix client",
     "build-backend": "npm run build --prefix server",
-    "build-shared": "npm run build --prefix shared",
     "build": "npm run build-frontend && npm run build-backend",
     "server": "npm run dev --prefix server",
     "client": "npm start --prefix client",
     "dev": "concurrently \"./wait-for-it.sh localhost:3306 -t 0 -- sleep 10 && npm run server\" \"npm run client\" \"docker-compose up \"",
-    "postinstall": "npm --prefix server install && npm --prefix client install && npm --prefix shared install",
-    "auditDep": "npm audit fix && npm --prefix server audit fix && npm --client server audit fix && npm --shared audit fix",
+    "postinstall": "npm --prefix server install && npm --prefix client install",
+    "auditDep": "npm audit fix && npm --prefix server audit fix && npm --client server audit fix",
     "prepare": "husky install",
     "test": "npm --prefix server test && CI=true npm --prefix client test -- --passWithNoTests",
-    "lint": "npm run --prefix server lint && npm run --prefix client lint && npm run --prefix shared lint",
-    "lint-ci": "npm run --prefix server lint-ci && npm run --prefix client lint-ci && npm run --prefix shared lint-ci",
+    "lint": "npm run --prefix server lint && npm run --prefix client lint",
+    "lint-ci": "npm run --prefix server lint-ci && npm run --prefix client lint-ci",
     "seq-cli": "npm run --prefix server seq-cli"
   },
   "author": "Open Government Products, GovTech Singapore (https://open.gov.sg)",
