+++ server/src/shared/types/base/common.ts
