+++ server/src/shared/types/base/index.ts
