+++ server/src/shared/types/base/permission.ts
