+++ server/src/shared/types/base/user.ts
