--- shared/.eslintrc.js
@@ -1,16 +1,0 @@
-module.exports = {
-  extends: [
-    'plugin:@typescript-eslint/recommended',
-    'plugin:prettier/recommended',
-  ],
-  rules: {
-    'no-console': 'warn',
-  },
-  parser: '@typescript-eslint/parser',
-  plugins: ['@typescript-eslint'],
-  ignorePatterns: ['node_modules/**/*'],
-  env: {
-    node: true,
-    es6: true,
-  },
-}
