--- shared/.lintstagedrc
@@ -1,3 +1,0 @@
-{
-  "*.{js,ts}": "eslint --fix"
-}
