--- shared/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "shared",
-  "version": "0.1.0",
-  "description": "Shared code between client and server for AskGovSG",
-  "main": "./build/index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "tsc",
-    "lint": "eslint . --quiet --fix --ext '.ts'",
-    "lint-ci": "eslint . --quiet --ext '.ts'",
-    "lint-staged": "lint-staged"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "zod": "^3.8.1"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.30.0",
-    "@typescript-eslint/parser": "^4.30.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "lint-staged": "^11.1.2",
-    "prettier": "^2.3.2",
-    "typescript": "^4.4.2"
-  }
-}
